adding IR sensor logic
the machine uses an IR sensor to determine if there's a cup or not.
The IR sensor is used to take n number of readings, and the avg reading
is compared to a preset threshold.  If there is a cup, it will pour,
but if not, it will hold onto the order and keep checking until a cup
is put down.